ultimas actualizaciones para final

+++ src/test/resources/data/h2db/schema.sql
@@ -1,3 +1,4 @@
+
 --
 -- Drop Table structure for table ventas_tarjeta
 --
@@ -101,12 +102,10 @@ CREATE TABLE ventas_efectivo (
 -- Table structure for table ventas_tarjeta
 --
 
-
-
 CREATE TABLE ventas_tarjeta (
   vta_id bigint NOT NULL,
   vtt_cantidad_cuotas int DEFAULT NULL,
-  vtt_coeficiente  decimal(2,2) DEFAULT NULL,
+  vtt_coeficiente  bigint DEFAULT NULL,
   PRIMARY KEY (vta_id),
   CONSTRAINT vtt_vta_fk FOREIGN KEY (vta_id) REFERENCES ventas (vta_id)
 );
